requirements.txt
pillow==11.0.0
py7zr==1.1.0
